requirements.txt
langchain
pypdf2
python-dotenv
streamlit
tiktoken
faiss-cpu